Show accurate* queued proposals count (#1204)
* Show accurate queued proposals count

This doesn't work on dev because the count from the subgraph doesn't exclude timed out proposals. On prod it should work do to the bot which executes timed out proposals.

* Show expired proposals in the scheme queued proposals list

So that the count of proposals matches what is shown. On prod of course we have a bot auto executing expired proposals so they still should disappear quickly.

* handle case where user is not logged in

* fix lint error

* fix actionbutton if balance is not known