Release 2019-12-10 (#1274)
* suppress exception on bad tokens (#1260)

* convert UNSAFE methods (#1261)

* shrink vertically, other cleanups (#1262)

* add subscriptions for votes, stakes and rewards (#1253)

* add subscriptions for votes, stakes and rewards

* only subscribe when on ProposalDetails page.

* do state too

* ignore externalTokenReward when address is null (#1265)

* change old dai to sai, add new dai (#1269)

* better position notifs on all pages (#1267)

* shorten token dropdown menu (#1271)

* add cookie policy and privacy policy (#1272)

* update links and text

* display PDFs in new tab

* dynamically compute url

* add PdfFile component, display all pages

* just to a window.redirect tp tje pdf file

* remove remainint react-pdf code

* 3box profiles and 3box feed (#1217)

* WIP almost finished with switch to 3box

* Remove comment, add account address to profile state

* Go back to not using profile-hover on dao members page

* Lint fix

* WIP

* Try to fix travis issue

* Another attempt to fix travis

* Fix travis??

* typoe on .travis.yml

* Fix build issues, using Terser and removing sourcemaps

* Load profiles from 3box for every AccountPopup and member on the Reputation Holders page

Can't use Box.getProfiles for now until 3box fixes it so have to load them one by one as needed in the app

* WIP store threeBox box in redux

* First pass at following items for the personal feed

Store followed items in a 3box space
Also use Redux so we can pass around what the current user is following throughout the app
Add UX to follow DAOs and Schemes so far

TODO: need a spinner and/or notification that the follow item is being saved because it can take a while
TODO: Add following of users and proposals

Add follow button to ProposalCard

Fix getting initial set of follows from 3box

* First version of feed page

Proposal specific feed item started

* Make it easier to size and style AccountPopup and AccountImage throughout the app

No longer need to pass in historyView, detailView... to the AccountPopup component, just a width, the outer component now positions it.

* Add following to proposal details page and account popup

Clean up styling of follow buttons throughout

* get feed data with a custom query

* Add timestamp to feed items

* Show user account images and names for feed items

Clean up styles and code for feed stuff in general

* Dao feed items

And proposal feed item descriptions

* Fix lint issues

* Fix route daoState data

* Give pointer cursor to scheme card follow button

* Fix styling of DAO feed item text

* Fix error on feed page

* Re-center login button in header

* Actually fix centering of login button in header

* Display proposal descriptions correctly.

Display first 300 characters for now
If more than that have a link to see the rest of the proposal
Display markdown

* Fix lint issue

* Add query for events by user, and dedupe all events

* Add UserFeedItem for NewReputationHolder events

* Remove ability to follow Schemes

* Make unconnected feed page look like designs and show all the DAOs

* Show user 3box profile icon in feed item title

* Show feed even if not connected to metamask account

* Prioritize showing events from followed proposals, then users, then DAOs

Meaning if the same event would show from following a proposal and a user and a DAO, show the proposal version first.

* Make feed the home page

Add empty feed banner when logged in but not following anything
Add All DAOs link to sidebar

* Fix tests with feed now as home page

* Fix follow button on Proposal Details Page

And fix profile reducer when unfollowing

* Fix lint error

* Fix display of linking too DAOs in feed item headers

* Add mini sidebar for non DAO pages like home page and all daos page

Move DaoSidebar to a shared SidebarMenu component that is now in AppContainer. Change the styling and width based on whether there is a DAO present or not.

TODO: make sure sidebar doesnt get covered when content width is smaller

* Center content until too small to fit with sidebar

* Fix lint issue

* Update to latest 3box 1.14

* Fix bugs on production when not following anything

* Create shared FollowButton component, and now with spinners!

* Fix feed icon image on production

* Fix strange merge bug?

* Clean up and small fixes

Make sure follow notification fits in the box
Add Tooltip to all DAOs link and fix one for home page
Remove commented out code and consoles

* Git pagination working on feed

This is kind of crappy and slow for now but ok for v1

* Remove unused code

* Catch errors when trying to follow something

So spinner doesnt spin forever on an error

* Fix lint error

* Ignore events with malformed JSON in the data field

https://github.com/daostack/subgraph/issues/406

* Fix image overflow in proposal descriptions in feed cards

* Show correct events with correct titles

Figured out a hacky way to filter events and dedupe correctly all on the client so it only shows the appropriate events for each followed entity type.
Also switched paging to be based on timestamp instead of the count of proposals because that makes a bit more sense. But still not great, and could miss some events...

* Add follow button to account profile page

* Fix lint issue

* update changelog for release 2019-12-10